pyproject.toml
[tool.poetry]
name = "evaproject"
version = "0.1.0"
description = ""
authors = ["Peter Kroeger <peter14mailgmail.com>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "1.21.6"
asgiref = "3.5.0"
astroid = "2.5.1"
asttokens = "2.0.4"
beautifulsoup4 = "4.9.3"
blinker = "1.4"
buttonshim = "0.0.2"
certifi = "^2022.12.7"
cffi = "1.15.1"
chardet = "4.0.0"
colorama = "0.4.6"
colorzero = "1.1"
comtypes = "^1.1.14"
cryptography = "^39.0.0"
distro = "^1.8.0"
distro-info = "^1.0"
django = "4.0.3"
docutils = "^0.19"
click = "7.1.2"
flask = "1.1.4"
fourletterphat = "^0.1.0"
fuzzywuzzy = "^0.18.0"
gpiozero = "^1.6.2"
html5lib = "^1.1"
idna = "^3.4"
imutils = "^0.5.4"
isort = "^5.11.4"
itsdangerous = "1.1.0"
jedi = "^0.18.2"
jinja2 = "2.11.3"
lazy-object-proxy = "1.9.0"
logilab-common = "^1.9.8"
lxml = "^4.9.2"
markupsafe = "1.1.1"
mccabe = "^0.7.0"
microdotphat = "^0.2.2"
mypy = "^0.991"
oauthlib = "^3.2.2"
mypy-extensions = "^0.4.3"
pantilthat = "^0.0.7"
parso = "^0.8.3"
olefile = "^0.46"
pexpect = "^4.8.0"
piglow = "^1.2.5"
pigpio = "^1.78"
pillow = "8.4.0"
pipenv = "^2022.12.19"
platformdirs = "^2.6.2"
psutil = "^5.9.4"
ptyprocess = "^0.7.0"
pycparser = "^2.21"
pygments = "^2.14.0"
pyjwt = "^2.6.0"
pyopenssl = "^23.0.0"
pyserial = "^3.5"
pysimplegui = "^4.60.4"
pytesseract = "^0.3.10"
python-dateutil = "^2.8.2"
pyttsx3 = "^2.90"
reportlab = "3.5.68"
requests = "^2.28.2"
requests-oauthlib = "^1.3.1"
roman = "2.0.0"
scrollphathd = "^1.3.0"
scrollphat = "^0.0.7"
send2trash = "^1.8.0"
sense-hat = "^2.4.0"
six = "^1.16.0"
skywriter = "^0.0.8"
sn3218 = "2.0.0"
sqlparse = "^0.4.3"
ssh-import-id = "^5.11"
toml = "^0.10.2"
typing-extensions = "^4.4.0"
urllib3 = "^1.26.14"
voice = "^0.1.0"
charset-normalizer = "^3.0.1"
cap1xxx = "^0.1.4"
blinkt = "^0.1.2"
dbus-python = "1.2.16"
deprecation = "^2.1.0"
drumhat = "^0.1.0"
automationhat = "^0.4.1"
envirophat = "^1.0.1"
explorerhat = "^0.5.1"
onboard-client = "^1.9.0"
packaging = "^23.0"
pgzero = "^1.2.1"
phatbeat = "^0.1.1"
pianohat = "^0.1.0"
psycopg2-binary = "^2.9.5"
pycairo = "^1.23.0"
pycurl = "^7.45.2"
pygobject = "^3.42.2"
pyinotify = "^0.9.6"
pyparsing = "^3.0.9"
rainbowhat = "^0.1.0"
rpi-gpio = "^0.7.1"
touchphat = "^0.0.2"
typed-ast = "^1.5.4"
unicornhathd = "^0.0.4"
pycups = "^2.0.1"
speechrecognition = "^3.9.0"
opencv-python = "4.4.0.46"
pysmbc = "^1.0.25.1"
python-levenshtein = "^0.20.9"
firebase-admin = "^6.1.0"
plotly = "5.13.0"
kaleido = "0.2.1"
pandas = "^1.5.3"
google-auth = "^2.16.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
